requirements.txt
pandas==1.1.5
PyPDF2==1.26.0
tabula-py==2.2.0
